// src/app/DTOs/UpdateCandidateDTO.DTOs.ts
export interface UpdateCandidateDTO {
    id: number;
    name: string;
    lastName: string;
    DOB: string;
    firstLineOfAddress: string;
    addressCity: string;
    addressCountry: string;
    addressPostCode: string;
    homePhoneNo: string;
    mobilePhoneNo: string;
    workPhoneNo: string;
    formattedDOB: string;
}